--- src/containers/ContributeTabContainer/ContributeTabContainerQueries.graphql
@@ -1,103 +1,0 @@
-query PrivilegesOnHubCollaboration($hubNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    collaboration {
-      ...PrivilegesOnCollaboration
-    }
-  }
-}
-query HubAspects($hubNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    collaboration {
-      id
-      callouts {
-        ...AspectsOnCallout
-        type
-      }
-    }
-  }
-}
-# challenge
-query PrivilegesOnChallengeCollaboration($hubNameId: UUID_NAMEID!, $challengeNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    challenge(ID: $challengeNameId) {
-      id
-      collaboration {
-        ...PrivilegesOnCollaboration
-      }
-    }
-  }
-}
-query ChallengeAspects($hubNameId: UUID_NAMEID!, $challengeNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    challenge(ID: $challengeNameId) {
-      id
-      collaboration {
-        id
-        callouts {
-          ...AspectsOnCallout
-          type
-        }
-      }
-    }
-  }
-}
-# opportunity
-query PrivilegesOnOpportunityCollaboration($hubNameId: UUID_NAMEID!, $opportunityNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    opportunity(ID: $opportunityNameId) {
-      id
-      collaboration {
-        ...PrivilegesOnCollaboration
-      }
-    }
-  }
-}
-query OpportunityAspects($hubNameId: UUID_NAMEID!, $opportunityNameId: UUID_NAMEID!) {
-  hub(ID: $hubNameId) {
-    id
-    opportunity(ID: $opportunityNameId) {
-      id
-      collaboration {
-        id
-        callouts {
-          ...AspectsOnCallout
-          type
-        }
-      }
-    }
-  }
-}
-
-fragment PrivilegesOnCollaboration on Collaboration {
-  id
-  authorization {
-    id
-    myPrivileges
-  }
-}
-
-mutation CreateAspectFromContributeTab($aspectData: CreateAspectOnCalloutInput!) {
-  createAspectOnCallout(aspectData: $aspectData) {
-    id
-    nameID
-    displayName
-    description
-    type
-    tagset {
-      id
-      name
-      tags
-    }
-    banner {
-      ...VisualUri
-    }
-    bannerNarrow {
-      ...VisualUri
-    }
-  }
-}
