+++ src/hooks/index.ts
@@ -22,7 +22,7 @@ export * from '../domain/hub/HubContext/useHub';
 export * from './useOrganization';
 export * from './useChallenge';
 export * from './useOpportunity';
-export * from './useAspectsData';
+export * from './useAspectCreatedOnCalloutSubscription';
 
 export * from './useUrlParams';
 export * from './useCurrentBreakpoint';
